Refactor wrapper into wrapper folder
